+++ parser/src/tests/asciidoc_lang/attributes/positional_and_named_attributes.rs
@@ -16,7 +16,7 @@ mod positional_attribute {
 
     use crate::{
         Parser,
-        blocks::{IsBlock, ListType, metadata::BlockMetadata},
+        blocks::{IsBlock, ListType, SimpleBlockStyle, metadata::BlockMetadata},
         content::SubstitutionGroup,
         tests::prelude::*,
     };
@@ -128,7 +128,15 @@ The markers are mapped as follows:
 
 Each shorthand entry is placed directly adjacent to previous one, starting immediately after the optional block style.
 The order of the entries does not matter, except for the style, which must come first.
-        
+The typical order is style, ID, role(s), option(s).
+
+All of these are valid:
+
+* `#idname.rolename`
+* `.rolename1.rolename2`
+* `.rolename#idname`
+* `#idname.rolename%optionname`
+
 "#
         );
 
@@ -427,6 +435,7 @@ Specifically, this syntax sets the ID to `rules`, adds the role `prominent`, and
                     col: 1,
                     offset: 31,
                 },
+                style: SimpleBlockStyle::Paragraph,
                 title_source: None,
                 title: None,
                 anchor: None,
@@ -500,6 +509,7 @@ Specifically, this syntax sets the `header`, `footer`, and `autowidth` options.
                     col: 1,
                     offset: 0,
                 },
+                style: SimpleBlockStyle::Paragraph,
                 title_source: None,
                 title: None,
                 anchor: None,
@@ -573,6 +583,7 @@ Formatted text does not support a style, so the first and only positional attrib
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -596,7 +607,7 @@ Formatted text does not support a style, so the first and only positional attrib
 mod named_attribute {
     use pretty_assertions_sorted::assert_eq;
 
-    use crate::{Parser, tests::prelude::*};
+    use crate::{Parser, blocks::SimpleBlockStyle, tests::prelude::*};
 
     non_normative!(
         r#"
@@ -653,6 +664,7 @@ A named attribute consists of a name and a value separated by an `=` character (
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -732,6 +744,7 @@ In all other cases, the surrounding quotes are optional.
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -812,6 +825,7 @@ If enclosing quotes are used, they are dropped from the parsed value and the pre
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -894,6 +908,7 @@ If enclosing quotes are used, they are dropped from the parsed value and the pre
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -930,7 +945,7 @@ If enclosing quotes are used, they are dropped from the parsed value and the pre
 mod attribute_list_parsing {
     use pretty_assertions_sorted::assert_eq;
 
-    use crate::{Parser, tests::prelude::*, warnings::WarningType};
+    use crate::{Parser, blocks::SimpleBlockStyle, tests::prelude::*, warnings::WarningType};
 
     non_normative!(
         r#"
@@ -991,6 +1006,7 @@ In these rules, `name` consists of a word character (letter or numeral) followed
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1096,6 +1112,7 @@ In these rules, `name` consists of a word character (letter or numeral) followed
                         col: 1,
                         offset: 27,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1182,6 +1199,7 @@ In these rules, `name` consists of a word character (letter or numeral) followed
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1246,6 +1264,7 @@ For subsequent attributes, any leading space or tab characters are skipped.
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1332,6 +1351,7 @@ Space and tab characters around the equals sign (=) and at the end of the value 
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1417,6 +1437,7 @@ Any space or tab characters at the boundaries of the value are ignored.
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1507,6 +1528,7 @@ Any space or tab characters at the boundaries of the value are ignored.
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1596,6 +1618,7 @@ Any space or tab characters at the boundaries of the value are ignored.
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1693,6 +1716,7 @@ If there is a closing double quote, the enclosing double quote characters are re
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1782,6 +1806,7 @@ If there is a closing double quote, the enclosing double quote characters are re
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1879,6 +1904,7 @@ If there is a closing single quote, the enclosing single quote characters are re
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,
@@ -1970,6 +1996,7 @@ If there is a closing single quote, and the first character is not an escaped si
                             col: 1,
                             offset: 47,
                         },
+                        style: SimpleBlockStyle::Paragraph,
                         title_source: None,
                         title: None,
                         anchor: None,
@@ -2075,6 +2102,7 @@ That's because the parser already knows to look for the closing square bracket a
                             col: 1,
                             offset: 41,
                         },
+                        style: SimpleBlockStyle::Paragraph,
                         title_source: None,
                         title: None,
                         anchor: None,
@@ -2175,6 +2203,7 @@ There are some exceptions to this rule, such as a link macro in a footnote, whic
                         col: 1,
                         offset: 0,
                     },
+                    style: SimpleBlockStyle::Paragraph,
                     title_source: None,
                     title: None,
                     anchor: None,